Move utilities functions to their own directory.
