+++ Main/main.cpp
@@ -16,8 +16,8 @@
 #include "DebugServer2/Host/Platform.h"
 #include "DebugServer2/Host/QueueChannel.h"
 #include "DebugServer2/Host/Socket.h"
-#include "DebugServer2/Log.h"
-#include "DebugServer2/OptParse.h"
+#include "DebugServer2/Utils/Log.h"
+#include "DebugServer2/Utils/OptParse.h"
 
 #include "SessionThread.h"
