+++ Sources/Architecture/ARM/SoftwareBreakpointManager.cpp
@@ -14,7 +14,7 @@
 #include "DebugServer2/Architecture/ARM/Branching.h"
 #include "DebugServer2/Target/Process.h"
 #include "DebugServer2/Target/Thread.h"
-#include "DebugServer2/Log.h"
+#include "DebugServer2/Utils/Log.h"
 
 #include <cstdlib>
