+++ Sources/GDBRemote/DummySessionDelegateImpl.cpp
@@ -11,7 +11,7 @@
 #include "DebugServer2/GDBRemote/DummySessionDelegateImpl.h"
 #include "DebugServer2/GDBRemote/Session.h"
 #include "DebugServer2/Host/Platform.h"
-#include "DebugServer2/Log.h"
+#include "DebugServer2/Utils/Log.h"
 
 using ds2::Host::Platform;
