+++ Sources/GDBRemote/Session.cpp
@@ -13,9 +13,9 @@
 #include "DebugServer2/GDBRemote/Session.h"
 #include "DebugServer2/GDBRemote/SessionDelegate.h"
 #include "DebugServer2/GDBRemote/ProtocolHelpers.h"
-#include "DebugServer2/SwapEndian.h"
-#include "DebugServer2/HexValues.h"
-#include "DebugServer2/Log.h"
+#include "DebugServer2/Utils/HexValues.h"
+#include "DebugServer2/Utils/Log.h"
+#include "DebugServer2/Utils/SwapEndian.h"
 
 #include <cstdlib>
 #include <cstring>
