+++ Sources/GDBRemote/SessionBase.cpp
@@ -14,8 +14,8 @@
 #include "DebugServer2/GDBRemote/SessionDelegate.h"
 #include "DebugServer2/GDBRemote/ProtocolHelpers.h"
 #include "DebugServer2/Host/Platform.h"
-#include "DebugServer2/HexValues.h"
-#include "DebugServer2/Log.h"
+#include "DebugServer2/Utils/HexValues.h"
+#include "DebugServer2/Utils/Log.h"
 
 #include <iomanip>
 #include <sstream>
