+++ Sources/Host/Linux/PTrace.cpp
@@ -13,7 +13,7 @@
 #include "DebugServer2/Host/Linux/PTrace.h"
 #include "DebugServer2/Host/Linux/ExtraWrappers.h"
 #include "DebugServer2/Host/POSIX/AsyncProcessWaiter.h"
-#include "DebugServer2/Log.h"
+#include "DebugServer2/Utils/Log.h"
 
 #include <cerrno>
 #include <csignal>
