+++ Sources/Host/POSIX/Platform.cpp
@@ -11,7 +11,7 @@
 #include "DebugServer2/Base.h"
 #include "DebugServer2/Host/Platform.h"
 #include "DebugServer2/Host/POSIX/Platform.h"
-#include "DebugServer2/Log.h"
+#include "DebugServer2/Utils/Log.h"
 
 #include <cstring>
 #include <fcntl.h>
