+++ Sources/Host/Windows/Platform.cpp
@@ -10,7 +10,7 @@
 
 #include "DebugServer2/Base.h"
 #include "DebugServer2/Host/Platform.h"
-#include "DebugServer2/Log.h"
+#include "DebugServer2/Utils/Log.h"
 #include "DebugServer2/Types.h"
 #include "DebugServer2/Host/Windows/ExtraWrappers.h"
